requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
pillow-heif